Use Qt paths for QML and plugins even with LIRI_INSTALL_PREFIX
Mimic the CMake and ECM behavior where there is a flag to install QML
modules and plugins to the Qt paths to make them available to apps
without setting the environment variables.

This mode is activated with `CONFIG+=use_qt_paths`.

In order to use Qt paths for everything (but headers)
use `CONFIG+=install_under_qt`.

Closes: #1